+++ bugsnag-android-core/src/main/java/com/bugsnag/android/Bugsnag.java
@@ -6,7 +6,10 @@ import android.content.Context;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * Static access to a Bugsnag Client, the easiest way to use Bugsnag in your Android app.
@@ -357,6 +360,23 @@ public final class Bugsnag {
      */
     public static void pauseSession() {
         getClient().pauseSession();
+    }
+
+    /**
+     * Returns the current buffer of breadcrumbs that will be sent with captured events. This
+     * ordered list represents the most recent breadcrumbs to be captured up to the limit
+     * set in {@link Configuration#getMaxBreadcrumbs()}.
+     *
+     * The returned collection is readonly and mutating the list will cause no effect on the
+     * Client's state. If you wish to alter the breadcrumbs collected by the Client then you should
+     * use {@link Configuration#setEnabledBreadcrumbTypes(Set)} and
+     * {@link Configuration#addOnBreadcrumb(OnBreadcrumbCallback)} instead.
+     *
+     * @return a list of collected breadcrumbs
+     */
+    @NonNull
+    public static List<Breadcrumb> getBreadcrumbs() {
+        return getClient().getBreadcrumbs();
     }
 
     /**